postgres/requirements.txt
huggingface_hub
pandas
pyarrow
sqlalchemy
python-dotenv
psycopg2-binary
